feat: add GetUser() to return current user (#343)
## Motivation

We are writing an internal API where we need to retrieve current user
for nobl9. Now, since it is not exposed outside the module, there was no
better way to do this. Someone else might need this as well.

## Summary

adds a `GetUser` func which returns user

## Release Notes

Added `GetUser` method for `Client` which returns the current session
user read from JWT token.

---------

Signed-off-by: shubhindia <shubhindia123@gmail.com>