feat: Add send resolution to Service Now (#703)
## Motivation

All-clear is not only supported for Pager Duty.
[[link]](https://docs.nobl9.com/alerting/alert-methods/pagerduty?_highlight=send#sending-notification-for-a-resolved-alert-to-pagerduty)
We want to enable it for ServiceNow as well. 

## Summary

- Add `SendResolution` field to `ServiceNowAlertMethod`, similarly to
`PagerDutyAlertMethod`
- Add validation max 128 characters to `sendResolution.message` (both
pagerduty and servicenow) to prevent from any `413` errors from alert
methods' APIs. There are no strict limits on the number of characters
but this is just a soft limit on our side.

## Release Notes

- `sendResolution` is now supported in Service Now alert method
configuration, which enables all-clear mechanism (closing alerts in
Service Now after notification cooldown period, with specified
`message`)