feat: Improve FileConfig handling (#729)
## Motivation

Currently it's hard to tell if file config has been loaded or not.
This PR changes the `GetFileConfig`contract to return a nil pointer if
no file config was loaded.
Previously, the function was returning only a shallow copy of
`FileConfig`, now it returns a proper deep copy.

## Breaking Changes

`sdk.Config.GetFileConfig` now returns a pointer.