feat: add support for label-based grouping in System Health Review Report [PC-17968] (#813)
## Summary

In addition to the added `rowGroupBy` enum value and new field for
`labelRows` and extended validation to cover these, I also improved the
test coverage a bit and added minor improvements to the validation logic
for Report.
On top of that the generated examples will now feature both types of
group by for SHR.

## Testing

Extended unit tests.

## Release Notes

v1alpha Report now has new groping mode for System Health Review. Set
`rowGroupBy = label` and configure `labelRows` to specify which label
should be used in the grouping.
Currently only a single label must be provided.