feat: remove JWK fetching and validation logic from JWT parser (#812)
## Motivation

Running a full JWT verification, including the signature through JWK is
a bit of an overkill on the client side and we've face some issues when
fetching the JWK (which is not cashed in config.toml) which affected the
clients reliability.
To improve the whole auth flow, we're removing full JWT verification and
instead focus on a few select claims.

## Tests

https://github.com/nobl9/nobl9-go/actions/runs/18678414140

## Release Notes

The `sdk.Client` no longer performs full JWT verification.