fix: Verify if the RemoveComputedFieldsFromObjects input value is a struct (#817)
## Motivation

If `sdk.RemoveComputedFieldsFromObjects` is called with
`v1alpha.GenericObject` it won't work since the `manifest.Object`
implementation is a generic map under the hood. In this scenario the
computed fields are not removed and the function fails silently. We
should harden validation for such scenarios.

## Breaking Changes

`sdk.RemoveComputedFieldsFromObjects` now returns an error.