feat: Add GetUserID and GetUserEmail methods to sdk.Client (#733)
## Motivation

To make things clearer `GetUser` is deprecated and renamed to
`GetUserEmail`.
In order to allow users to access the user ID directly `GetUserID` is
being added.

## Testing

https://github.com/nobl9/nobl9-go/actions/runs/19033360284

## Release Notes

Added `sdk.Client.GetUserID` which returns ID of the user associated
with the access token used by `sdk.Client`.
Added `sdk.Client.GetUserEmail` which is a rename of
`sdk.Client.GetUser` function.

## Breaking Changes

`sdk.Client.GetUser` is now deprecated and will be removed in the
future. Use `sdk.Client.GetUserEmail` instead.

---------

Co-authored-by: daniel-zelazny <83642555+daniel-zelazny@users.noreply.github.com>