chore: migrate to golangci lint v2 and switch to gofumpt for formatting Go files (#828)
## Motivation

Golangci lint v2 comes with `fmt` command which can be used to format
all files according to rules defined by listed tools.
This way we don't have to manage multiple binaries but can solely rely
on `golangci-lint` to do the job and verify if the files were formatted.

### gofumpt

`gofumpt` comes with a number of sensible formatting rules which above
all improve readability of the codebase.
It further reduces the mental overhead of worrying about formatting
rules.