feat: add ProjectScopedKinds() function (#856)
## Motivation

Currently, in order to get all object kinds which are project-scoped one
must either define it manually (requires manual maintenance) or define a
list of all the objects' structs, iterate over it and check if they
implement `ProjectScopedObject` interface.
This is tedious, ideally we'd want to have a simple function that
returns a slice of all `manifest.Kind` representing all project-scoped
objects.

## Summary

This PR introduces the `kindgen` tool, which automates the generation of
the `ProjectScopedKinds()` function in the `manifest` package.

The tool scans `*_object.go` files in the `manifest/v1alpha/` directory
to identify types implementing the `ProjectScopedObject` interface. It
then maps these types to their corresponding `Kind` constants and
generates the `kind_project_scoped.go` file. The generated function
returns a list of all project-scoped kinds.

The `kindgen` tool is integrated into the `make generate/code` process
to ensure deterministic execution order. A detailed README is included
to document its usage and functionality.

## Release Notes

You can now call `manifest.ProjectScopedKinds()` function to get a list
of `manifest.Kind` for all objects which are project-scoped.